uk-esos-app-api/src/test/java/uk/gov/esos/api/reporting/noc/phase3/validation/NocP3ContactPersonsContextValidatorServiceTest.java
package uk.gov.esos.api.reporting.noc.phase3.validation;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import uk.gov.esos.api.reporting.noc.common.domain.NocValidationResult;
import uk.gov.esos.api.reporting.noc.common.domain.NocViolation;
import uk.gov.esos.api.reporting.noc.common.validation.NocSectionConstraintValidatorService;
import uk.gov.esos.api.reporting.noc.phase3.domain.NocP3;
import uk.gov.esos.api.reporting.noc.phase3.domain.NocP3Container;
import uk.gov.esos.api.reporting.noc.phase3.domain.ReportingObligationCategory;
import uk.gov.esos.api.reporting.noc.phase3.domain.contactpersons.ContactPersons;

import java.util.Optional;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.verifyNoInteractions;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class NocP3ContactPersonsContextValidatorServiceTest {

    @InjectMocks
    private NocP3ContactPersonsContextValidatorService contextValidator;

    @Mock
    private NocSectionConstraintValidatorService<ContactPersons> nocSectionConstraintValidatorService;

    @Test
    void validate() {
        final ContactPersons contactPersons = ContactPersons.builder().hasSecondaryContact(false).build();
        final NocP3Container nocContainer = NocP3Container.builder()
                .noc(NocP3.builder().contactPersons(contactPersons).build())
                .build();

        when(nocSectionConstraintValidatorService.validate(contactPersons)).thenReturn(Optional.empty());

        // Invoke
        NocValidationResult result = contextValidator
                .validate(nocContainer, ReportingObligationCategory.ALTERNATIVE_ENERGY_ASSESSMENTS_95_TO_100);

        // Verify
        assertThat(result.isValid()).isTrue();

        verify(nocSectionConstraintValidatorService, times(1)).validate(contactPersons);
    }

    @Test
    void validate_not_valid_for_category() {
        final ContactPersons contactPersons = ContactPersons.builder().hasSecondaryContact(false).build();
        final NocP3Container nocContainer = NocP3Container.builder()
                .noc(NocP3.builder().contactPersons(contactPersons).build())
                .build();

        // Invoke
        NocValidationResult result = contextValidator
                .validate(nocContainer, ReportingObligationCategory.NOT_QUALIFY);

        // Verify
        assertThat(result.isValid()).isFalse();
        assertThat(result.getNocViolations()).extracting(NocViolation::getMessage)
                .containsOnly(NocViolation.NocViolationMessage.INVALID_SECTION.getMessage());
        assertThat(result.getNocViolations()).extracting(NocViolation::getSectionName)
                .containsOnly(ContactPersons.class.getName());

        verifyNoInteractions(nocSectionConstraintValidatorService);
    }

    @Test
    void getApplicableReportingObligationCategories() {
        assertThat(contextValidator.getApplicableReportingObligationCategories()).containsExactlyInAnyOrder(
                ReportingObligationCategory.ESOS_ENERGY_ASSESSMENTS_95_TO_100,
                ReportingObligationCategory.ISO_50001_COVERING_ENERGY_USAGE,
                ReportingObligationCategory.PARTIAL_ENERGY_ASSESSMENTS,
                ReportingObligationCategory.ZERO_ENERGY,
                ReportingObligationCategory.LESS_THAN_40000_KWH_PER_YEAR,
                ReportingObligationCategory.ALTERNATIVE_ENERGY_ASSESSMENTS_95_TO_100
        );
    }
}
